Optimized for small screen
The navigation bar kept overlaying the content. I created a div with a class of content-container which I placed directly under the navigation bar and just before the h2 and h1 of the index and about page respectively.

I then added a padding top and this solved the problem of the navigation bar overlaying the content on small screen